+++ Assets/Scripts/Core/Pool/IPooler.cs.meta
