--- Assets/Scripts/Core/Spawner/RandomizerPooler.cs
@@ -1,20 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class RandomizerPooler : Pooler
-{
-    private readonly IRandomizer _randomizer;
-        
-    public RandomizerPooler(IList<GameObject> poolingObjects, IRandomizer randomizer) : base(poolingObjects)
-    {
-        _randomizer = randomizer;
-    }
-        
-    protected override int GetIndex(int maxValue)
-    {
-        _randomizer.SetMax(maxValue);
-            
-        return _randomizer.GetIndex();
-    }
-}
