+++ Assets/Scripts/Data/IterationPatternsData.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace Data
+{
+    [CreateAssetMenu]
+    public class IterationPatternsData : ScriptableObject
+    {
+        [SerializeField] private PatternsData[] _iterationPatternsData;
+
+        public PatternsData[] IterationPatterns => _iterationPatternsData;
+    }
+}
