--- Assets/Scripts/Data/LevelPatternsData.cs
@@ -1,12 +1,0 @@
-﻿using UnityEngine;
-
-namespace Data
-{
-    [CreateAssetMenu]
-    public class LevelPatternsData : ScriptableObject
-    {
-        [SerializeField] private PatternsData[] _levelPatternsData;
-
-        public PatternsData[] LevelPatterns => _levelPatternsData;
-    }
-}
