+++ Assets/Scripts/GameLifetimeScope.cs
@@ -21,7 +21,7 @@ public class GameLifetimeScope : LifetimeScope
     [SerializeField] private ItemsData _itemsData;
     [SerializeField] private LevelData _levelData;
     [SerializeField] private ItemsOffsetData _itemsOffsetData;
-    [SerializeField] private LevelPatternsData _levelPatternsData;
+    [SerializeField] private IterationPatternsData _iterationPatternsData;
 
     protected override void Configure(IContainerBuilder builder)
     {
@@ -30,7 +30,7 @@ public class GameLifetimeScope : LifetimeScope
         BindInstance(builder, _itemsData);
         BindInstance(builder, _levelData);
         BindInstance(builder, _itemsOffsetData);
-        BindInstance(builder, _levelPatternsData);
+        BindInstance(builder, _iterationPatternsData);
     }
 
     private void BindComponents(IContainerBuilder builder)
@@ -52,7 +52,7 @@ public class GameLifetimeScope : LifetimeScope
         builder.Register<MainCameraBorders>(Lifetime.Singleton).AsSelf();
         builder.Register<PatternModel>(Lifetime.Singleton).AsSelf();
         builder.Register<SliderController>(Lifetime.Singleton).AsSelf();
-        builder.Register<LevelPatternsModel>(Lifetime.Singleton).AsSelf();
+        builder.Register<IterationModel>(Lifetime.Singleton).AsSelf();
         builder.Register<LevelController>(Lifetime.Singleton).AsSelf();
         builder.Register<ItemController>(Lifetime.Singleton).AsSelf(); // Как можно передвигать строки с помощью alt как в VS
